refactor(packages): shared/core の境界明確化 + Web/Mobile 重複統合
- packages/shared に date-utils.ts (formatLocalDate, todayLocal, parseLocalDate, addDays) を追加
- packages/shared に nutrition-planner.ts (deriveMacroTargets, estimateGoalProjection) を追加
- src/lib/date-utils.ts を @homegohan/shared からの re-export に変更
- src/lib/nutrition-target-planner.ts を @homegohan/shared からの re-export に変更
- src/hooks/useHomeData.ts, meals/new/page.tsx, menus/weekly/page.tsx のインライン formatLocalDate を削除し @homegohan/shared に統一
- apps/mobile/src/lib/mealPlan.ts のインライン formatLocalDate を @homegohan/shared に統一
- apps/mobile/app/onboarding/complete/index.tsx のインライン deriveMacros を @homegohan/shared の deriveMacroTargets に置換
- apps/mobile/tsconfig.json に @homegohan/core パスエイリアスを追加
- root tsconfig.json に @homegohan/core パスエイリアスを追加
- packages/shared/README.md, packages/core/README.md に境界定義を明記
