+++ apps/mobile/src/lib/mealPlan.ts
@@ -1,11 +1,5 @@
 import { supabase } from "./supabase";
-
-const formatLocalDate = (date: Date): string => {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, "0");
-  const d = String(date.getDate()).padStart(2, "0");
-  return `${y}-${m}-${d}`;
-};
+import { formatLocalDate } from "@homegohan/shared";
 
 /**
  * Get or create user_daily_meals for a specific date
