+++ packages/shared/src/index.ts
@@ -4,3 +4,5 @@ export * from './dish-roles';
 export * from './meal-types';
 export * from './mode-config';
 export * from './progress-phases';
+export * from './date-utils';
+export * from './nutrition-planner';
